clock: fix uint16_t overflow in brightness moving-average sum
update_average_brightness() accumulated the broadband samples into a uint16_t.
broadband is tiny in normal light (dim room ~1) but reaches tens of thousands
under a bright source, so summing moving_ave_period of them wrapped past 65535
and produced a garbage average that swung ~200<->3500 sample-to-sample.

The old ~55 s ramp masked this (it low-pass-filtered the garbage target away).
The new ~5 s ramp follows the target closely, so the corrupted average made the
display lurch down-up-down while dimming after a torch was removed. Brightening
looked fine because the target was pinned at the 4015 clamp the whole time.

Widen the accumulator to uint32_t (holds 10 * 65535 comfortably). Latent for
years because ordinary ambient never summed high enough to overflow.

Verified on-device: rebuilt (gcc-14, -Werror) and redeployed; the average is a
true 10-sample mean again. Torch down-ramp re-test pending.

+++ clock.c
@@ -215,7 +215,13 @@ uint16_t brightness_map(float brightness) {
 
 void update_average_brightness(void) {
 
-  uint16_t sum = 0;
+  // sum must be wide enough for moving_ave_period broadband samples. broadband
+  // is small in normal light but reaches tens of thousands under bright light
+  // (e.g. a torch on the sensor); a uint16_t sum wrapped past 65535 there,
+  // producing a garbage average that swung ~200<->3500 sample-to-sample and
+  // made the ramp lurch on the way down. 32 bits holds 10 * 65535 with room to
+  // spare. (Latent for years: normal ambient never summed high enough.)
+  uint32_t sum = 0;
   for (uint16_t x = 0; x < brightness_samples ; x++) {
     sum = sum + brightness_buffer[x];
   }
